fixing annotation

+++ 0x02-python_async_comprehension/1-async_comprehension.py
@@ -5,6 +5,6 @@ from typing import List
 async_generator = __import__('0-async_generator').async_generator
 
 
-async def async_comprehension() -> List[int]:
+async def async_comprehension() -> List[float]:
     """ async comprehension that returns list of floats"""
     return [a async for a in async_generator()]
